backend/requirements_org.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

# azure-functions
# azure.storage
# networkx
# functools
# six
# python-jose-cryptodome

# azure-storage-file-datalake
# numpy==1.19.5
# pandas==1.1.4
# simplejson
# pyodbc
# pycaret==2.3.0
# category_encoders
# fsspec
# xgboost==1.2.0

# azure-storage-file-datalake==12.6.0
# numpy==1.19.5
# pandas==1.1.4
# simplejson==3.17.6
# pycaret==2.3.0
# fsspec==2022.2.0
# xgboost==1.2.0
# category-encoders==2.4.0


# aiohttp
# pycaret[FULL]==2.3.0
# scipy
# xlrd
# matplotlib
# statsmodels
# sklearn
# yellowbrick
# requests==2.19.1

# azure-common==1.1.27
# azure-functions==1.7.0
# azure-nspkg==3.0.2
# azure-storage==0.36.0
# azure-storage-blob==1.4.0
# azure-storage-common==1.4.2
# networkx==2.5.1
# six==1.15.0
# python-jose-cryptodome==1.3.2
# pyodbc==4.0.30
# numpy==1.18.2
# pandas==1.0.3
# scipy==1.2.1
# xlrd==1.2.0
# matplotlib==3.2.1
# statsmodels==0.11.1


aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.7
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
autopage==0.5.0
autopep8==1.6.0
awscli==1.22.80
azure-core==1.23.0
azure-storage-blob==12.10.0
azure-storage-file-datalake==12.6.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
blis==0.7.7
Boruta==0.3
botocore==1.24.25
cachetools==5.0.0
catalogue==1.0.0
catboost==1.0.4
category-encoders==2.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
cliff==3.10.1
cloudpickle==2.0.0
cmaes==0.8.2
cmd2==2.4.0
colorama==0.4.4
colorlog==6.6.0
colorlover==0.3.0
cryptography==36.0.2
cufflinks==0.17.3
cycler==0.11.0
cymem==2.0.6
Cython==0.29.23
databricks-cli==0.16.4
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.13
django-tinymce==3.4.0
docker==5.0.3
docutils==0.15.2
ecdsa==0.17.0
entrypoints==0.4
et-xmlfile==1.1.0
fastapi==0.75.2
filelock==3.6.0
Flask==2.0.3
Flask-Cors==3.0.10
fonttools==4.30.0
frozenlist==1.3.0
fsspec==2022.2.0
funcy==1.17
future==0.18.2
gensim==4.1.2
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.7.1
google-auth==2.6.2
google-cloud-core==2.2.3
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
graphviz==0.19.1
greenlet==1.1.2
grpcio==1.43.0
gunicorn==20.1.0
h11==0.13.0
htmlmin==0.1.12
httptools==0.4.0
hyperopt==0.2.7
idna==3.3
ImageHash==4.2.1
imbalanced-learn==0.9.0
importlib-metadata==4.11.3
importlib-resources==5.4.0
ipykernel==6.9.2
ipython==7.32.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.0
jmespath==1.0.0
joblib==1.0.1
jose==1.0.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.0
kmodes==0.11.1
lightgbm==3.3.2
llvmlite==0.38.0
Mako==1.2.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
missingno==0.5.1
mistune==0.8.4
mlflow==1.24.0
mlxtend==0.19.0
msgpack==1.0.3
msrest==0.6.21
multidict==6.0.2
multimethod==1.7
murmurhash==1.0.6
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
networkx==2.6.3
nltk==3.7
notebook==6.4.10
numba==0.55.1
numexpr==2.8.1
numpy==1.19.5
oauthlib==3.2.0
openpyxl==3.0.10
optuna==2.10.0
packaging==21.3
pandas==1.1.4
pandas-profiling==3.1.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pbr==5.8.1
phik==0.12.2
pickleshare==0.7.5
Pillow==9.0.1
plac==1.1.3
plotly==5.6.0
preshed==3.0.6
prettytable==3.2.0
prometheus-client==0.13.1
prometheus-flask-exporter==0.19.0
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
psycopg2==2.9.3
py4j==0.10.9.5
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycaret==2.3.0
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.3.1
pydantic==1.9.0
Pygments==2.11.2
pyLDAvis==3.3.1
pynndescent==0.5.6
pyod==0.9.8
pyodbc==4.0.32
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyperclip==1.8.2
pyreadline3==3.4.1
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-jose-cryptodome==1.3.2
pytz==2021.3
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
querystring-parser==1.2.4
ray==1.11.0
redis==4.2.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.5.2
scikit-learn==0.23.2
scikit-optimize==0.9.0
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
shap==0.40.0
simplejson==3.17.6
six==1.16.0
sklearn==0.0
slicer==0.0.7
smart-open==5.2.1
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.1
spacy==2.3.7
SQLAlchemy==1.4.32
sqlparse==0.4.2
srsly==1.0.5
starlette==0.17.1
statsmodels==0.13.2
stevedore==3.5.0
tabulate==0.8.9
tangled-up-in-unicode==0.1.0
tenacity==8.0.1
tensorboardX==2.5
terminado==0.13.3
testpath==0.6.0
textblob==0.17.1
thinc==7.4.5
threadpoolctl==3.1.0
toml==0.10.2
tornado==6.1
tqdm==4.63.1
traitlets==5.1.1
tune-sklearn==0.4.1
typing_extensions==4.1.1
umap-learn==0.5.2
urllib3==1.23
uvicorn==0.17.6
visions==0.7.4
waitress==2.1.1
wasabi==0.9.0
watchgod==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
websockets==10.3
Werkzeug==2.0.3
widgetsnbextension==3.6.0
wordcloud==1.8.1
wrapt==1.14.0
xgboost==1.2.0
xlrd==2.0.1
yarl==1.7.2
yellowbrick==1.4
zipp==3.7.0